tried F.cpp to 'F'

+++ 2017/0603_ARC075/F.cpp
@@ -40,7 +40,7 @@ ll rev(ll n) {
     s = '0' + s;
   }
   reverse(s.begin(), s.end());
-  return stoi(s);
+  return stoll(s);
 }
 
 int main () {
